refactor: move api helpers to separate file
